% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox_inverse}
\alias{boxcox_inverse}
\title{Reverse Box-Cox Transformation}
\usage{
boxcox_inverse(x, lambda, shift)
}
\arguments{
\item{x}{Vector that was modified by the Box-Cox transformation}

\item{lambda}{Value for the lambda parameter of the Box-Cox transformation}

\item{shift}{Shift amount from Box-Cox transformation}
}
\value{
Data x transformed to the original scale before the Box-Cox
transformation
}
\description{
Compute the reverse Box-Cox transformation for improvement
of the Normality and residual assumptions.
}
\examples{
data <- data.frame('value'=c(rnorm(14,sd = 2), rnorm(6), rnorm(20,mean=2)),
                   'group'=c(rep('A',14),rep('B',6), rep('C',20)))
tmp <- boxcox_transformation(data)
boxcox_inverse(tmp$data[,1],tmp$lambda, tmp$shift)
}
