% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox_transformation}
\alias{boxcox_transformation}
\title{Box-Cox Transformation}
\usage{
boxcox_transformation(data, lambdas = seq(-3, 3, 1/10))
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}

\item{lambdas}{Numeric values for potential lamba value}
}
\value{
List with shift data, lambda parameter, and shift amount to ensure
no negative values
}
\description{
Compute the Box-Cox transformation for improvement
of the Normality and residual assumptions.
}
\examples{
data <- data.frame(
  "value" = c(rnorm(14, sd = 2), rnorm(6), rnorm(20, mean = 2)),
  "group" = c(rep("A", 14), rep("B", 6), rep("C", 20))
)
boxcox_transformation(data)
}
