% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot_strata}
\alias{boxplot_strata}
\title{Title}
\usage{
boxplot_strata(data, strata)
}
\arguments{
\item{data}{Data.frame of the data, where rows are observations and columns
are the variables.}

\item{strata}{String of the column name in the data to investigate.}
}
\value{
A ggplot2 object of strata sizes
}
\description{
Title
}
\examples{
boxplot_strata(data.frame('A'=rnorm(100),'B'=rbinom(100,2,0.5)), 'B')
}
