% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{credit}
\alias{credit}
\title{German Credit Data}
\format{
\subsection{\code{credit}}{

A data frame with 1000 rows and 14 columns:
\describe{
\item{Class}{Credit worthiness as 'Good' or Bad'.}
\item{Age}{Age of person.}
\item{NumberExistingCredits}{Number of existing credits. }
\item{Personal}{Sex and relationship status of person. }
\item{CheckingAccountStatus}{Status of the checking account. }
\item{CreditHistory}{Status of credit history. }
\item{Purpose}{Reason for requesting credit. }
\item{SavingsAccountBonds}{ Amount of Savings account bonds. }
\item{EmploymentDuration}{Length of current employment. }
\item{OtherDebtorsGuarantors}{Information on if there is a guarantor. }
\item{Property}{Type of property. }
\item{OtherInstallmentPlans}{Information on other installment plans.}
\item{Housing}{Status of thier house. }
\item{Job}{Type of job. }
}
}
}
\source{
UCI Machine Learning Repository
}
\usage{
credit
}
\description{
Data on credit worthiness.
}
\references{
Kuhn, M. (2008). Building Predictive Models in R Using the caret
Package. Journal of Statistical Software, 28(5), 1–26.
https://doi.org/10.18637/jss.v028.i05
}
\keyword{datasets}
