% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plot.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Distribution Plot}
\usage{
distribution_plot(
  x,
  distribution = "norm",
  alpha = 0.05,
  labels = names(x),
  ...
)
}
\arguments{
\item{x}{Vector of observations}

\item{distribution}{String of distribution as part of the 'qDIST' and 'dDIST'}

\item{alpha}{Numeric for the significance used for the bands. If it outside
(0,1) or NULL then no bands will be computed}

\item{labels}{Vector of labels of the same length as \code{x} for the string to
add to the plot when a value exceeds the bounds}

\item{...}{Additional parameters to pass to qDIST}
}
\value{
A ggplot2 object for the values plotted against theoretical values
}
\description{
Plot data against a theoretical distribution. The typical comparision to a
normal distribution is called a QQ-plot.
}
\examples{
distribution_plot(rnorm(100))
distribution_plot(rexp(100), distribution = "exp")
}
