% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{.convertSamplesToDF}
\alias{.convertSamplesToDF}
\title{Convert List of Samples into a Data Frame}
\usage{
.convertSamplesToDF(data_list)
}
\arguments{
\item{data_list}{List of elements to be combined to a data.frame.}
}
\value{
Data.frame of the data in data_list
}
\description{
This (internal) function takes a list with differ length data.frames or
vectors and pads them all to make a clean data.frame.
}
\details{
This is an internal function and will not be viewable to user. See
generalized_resampling for usage
}
\keyword{internal}
