% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{.getChunks}
\alias{.getChunks}
\title{Get Chunks}
\usage{
.getChunks(x, chunksN)
}
\arguments{
\item{x}{Vector of values}

\item{chunksN}{Numeric indicating the number of chunks to split data into}
}
\value{
A list with chunksN items, each containing an similiar sized subset
of the original vector
}
\description{
This (internal) function splits the vector x into a \code{chunksN} number of
subsegments. The values of x are kept in order (i.e. not scrambled).
}
\examples{
.getChunks(1:100, 1)
.getChunks(1:100, 2)
.getChunks(1:100, 5)
}
\keyword{internal}
