% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.PMCMRplus_summary}
\alias{.PMCMRplus_summary}
\title{Summarize Results From PMCMRplus Without Outputs}
\usage{
.PMCMRplus_summary(x, alpha)
}
\arguments{
\item{x}{PMCMRplus data}

\item{alpha}{significance in [0, 1]}
}
\value{
Summary data.frame
}
\description{
Summarize Results From PMCMRplus Without Outputs
}
