% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_data}
\alias{.prepare_data}
\title{Prepare Data Utility}
\usage{
.prepare_data(data)
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}
}
\value{
List with data, groups, and formula
}
\description{
Prepare Data Utility
}
