% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{eta_squared}
\alias{eta_squared}
\title{Eta Squared}
\usage{
eta_squared(data)
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}
}
\value{
Table with eta-squared attached to classic ANOVA decomposition
}
\description{
Compute the Eta squared effect size statistic.
}
\examples{
data <- data.frame(
  "value" = c(rnorm(14, sd = 2), rnorm(6), rnorm(20, mean = 2)),
  "group" = c(rep("A", 14), rep("B", 6), rep("C", 20))
)
eta_squared(data)
}
\references{
Navarro, D. J. (2015) Learning statistics with R: A tutorial for
psychology students and other beginners. (Version 0.6) University of New
South Wales. Sydney, Australia
}
