% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_howell.R
\name{games_howell}
\alias{games_howell}
\title{Games-Howell Test}
\usage{
games_howell(data, alpha = 0.05)
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}

\item{alpha}{Significance for confidence intervals, defaults to 0.05}
}
\value{
Table showing groups, their differences, and significance, among other
details
}
\description{
Computes Games-Howell test on data. Similar to Tukey HSD, it computes the
significance of the differences between groups. However, it requires far
fewer assumptions. Per the original paper, non-normality it not a problems,
groups can be different sizes, and there is assumption on the homogeneity of
variances. Further, small sample sizes are okay (each group is typically
recommended to have at least 6 observations).
}
\examples{
data <- data.frame('value'=c(rnorm(14,sd = 2), rnorm(6), rnorm(20,mean=2)),
                   'group'=c(rep('A',14),rep('B',6), rep('C',20)))
games_howell(data)
}
\references{
Games, P. A., & Howell, J. F. (1976). Pairwise Multiple
Comparison Procedures with Unequal N’s and/or Variances: A Monte Carlo
Study. Journal of Educational Statistics, 1(2), 113–125.

Games, P. A., Keselman, H. J., & Clinch, J. J. (1979). Tests for
homogeneity of variance in factorial designs. Psychological Bulletin, 86(5),
978–984.
}
\seealso{
\code{\link[stats:TukeyHSD]{stats::TukeyHSD()}}
}
