% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinomial}
\alias{multinomial}
\alias{pmultinomial}
\alias{dmultinomial}
\alias{rmultinomial}
\title{Multinomial Distribution}
\usage{
pmultinomial(qs, probs, size = sum(qs))

dmultinomial(xs, probs)

rmultinomial(n, size, probs)
}
\arguments{
\item{qs}{Numeric vector for the max number of successes in each class}

\item{probs}{Numeric vector for the probabilities of each class}

\item{size}{Integer for the total number of successes accross all classes}

\item{xs}{Numeric vector for number of each class that is a success}

\item{n}{Numeric for number of random vectors to draw}
}
\value{
Probability that no more than \code{qs} successes are observed

Probability of \code{xs} successes under \code{probs} in a multinomial
distribution

Randomly observed multinomial distribution. Results are a matrix
with each column indicating a separate trial and the rows for each possible
outcome
}
\description{
Generate multinomially distributed random number vectors and compute
multinomial probabilities.
}
\examples{
pmultinomial(c(2,2,2),c(1/3,1/3,1/3),6)
pmultinomial(c(3,1,2),c(1/2,1/4,1/4),4)
dmultinomial(c(1,2,3),c(1,2,3))
dmultinomial(c(1,2,3),c(1/3,1/3,1/3))
rmultinomial(10, 5, c(1/2,1/4,1/4))
rmultinomial(5, 8, c(10,2,4))
}
