% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_probability.R
\name{observation_probability}
\alias{observation_probability}
\title{Observation Probability}
\usage{
observation_probability(strata_info)
}
\arguments{
\item{strata_info}{Data.frame with the first column having the number of
observed occurrences and the second column being the probability of the
events}
}
\value{
Numeric probability in [0,1] indicating the probability of an equal
to or more extreme distribution of events
}
\description{
Evaluates the probability that randomly drawing from the groups with given
probabilities would return a distribution of results the same or more
extreme (i.e. less likely). Since the same or more extreme outcomes always
include the given outcome, this probability will always be equal to or larger
than the probability of the given event.
}
\examples{
observation_probability(
  data.frame("counts" = c(15, 5), c(0.5, 0.5))
)
observation_probability(
  data.frame("counts" = c(15, 5), c(0.9, 0.1))
)
observation_probability(
  data.frame("counts" = c(10, 10), c(0.5, 0.5))
)
}
