% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_strata_bar}
\alias{plot_strata_bar}
\title{Plot Barplot for Strata Sizes}
\usage{
plot_strata_bar(data)
}
\arguments{
\item{data}{Vector of data with values of strata.}
}
\value{
A ggplot2 object of barplots for strata counts.
}
\description{
Produce a barplot of strata sizes based on data.
}
\examples{
data <- data.frame('A'=rnorm(100),'B'=rbinom(100,2,0.5))
plot_strata_bar(data$B)
}
