% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_strata_box}
\alias{plot_strata_box}
\title{Plot Boxplot for Strata Sizes}
\usage{
plot_strata_box(data)
}
\arguments{
\item{data}{Data.frame of the data, the first column is the strata and the
second some value for the box plots.}
}
\value{
A ggplot2 object of boxplots for data based on strata.
}
\description{
Produce a boxplot of strata sizes based on data.
}
\examples{
plot_strata_box(data.frame('B'=rbinom(100,2,0.5),'A'=rnorm(100)))
}
