% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{resample_function}
\alias{resample_function}
\title{Resample and apply data}
\usage{
resample_function(data, fn, alpha = 0.05, ...)
}
\arguments{
\item{data}{Data.frame of to be resampled where the rows are the observations
and the columns are the variables. Note, all variables except the first are
permuted, where the first is used for the models given in fn.}

\item{fn}{Function to apply on each resampled data set. It should return
a data,frame with the first column at the variable estimates (row as
variables).}

\item{...}{This is for the parameters of the resampling. See parameters of
\code{resample()}, if not given, the defaults are used.}
}
\value{
A list of items. The first is the results as a dataframe. It has the
following columns with a row per variable.
\itemize{
\item estimates: Mean of the parameter estimates from fn over the
resampled trials.
\item sd: Standard deviations of the estimates from \code{fn} over the
resampled trials.
\item lowerNorm: Lower confidence interval assuming normality using alpha.
\item upperNorm: Upper confidence interval assuming normality using alpha.
\item lowerResamp: Lower confidence interval from data using alpha.
\item upperResamp: Upper confidence interval from data using alpha.
}
The second entry is a list of the resampled data and estimates for reference.
\itemize{
\item samples: A list of the resampled data.
\item function_values: A list of the function output values from the
resampled data. Note, this includes those not used for the estimates.
\item estimates: A data frame of the first column of each element in
function_values. The row means are those used in \code{meanParams}.
}
}
\description{
Resample data and compute the related values from some function. This
function is a convenient an extension of \code{resample()} and a similar
function can be constructed using \code{resample()} when
\code{resample_function()} is insufficient. The code permutes every
column except the first ('output'), then applies the function on
the permuted sample with the (non-permuted) output.
}
