% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tests.R
\name{resample_welch_anova}
\alias{resample_welch_anova}
\title{Resample Welch ANOVA}
\usage{
resample_welch_anova(data, var.equal = FALSE, M = 1000)
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}

\item{var.equal}{Boolean if we can assume equal variance over the groups}

\item{M}{Numeric for the number of iterations in the resampling}
}
\value{
p-value for the effect of the groups
}
\description{
Resample data for significance of groups when running an ANOVA \code{\link[stats:oneway.test]{stats::oneway.test()}}
}
\examples{
data <- data.frame(
  "value" = c(rnorm(14, sd = 2), rnorm(6), rnorm(20, mean = 2)),
  "group" = c(rep("A", 14), rep("B", 6), rep("C", 20))
)
resample_welch_anova(data, M = 100)
}
