% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{subcats}
\alias{subcats}
\title{Subset of the Anatomical Data from Domestic Cats}
\format{
\subsection{\code{subcats}}{

A data frame with 116 rows and 3 columns:
\describe{
\item{Sex}{Sex of cat as a factor with levels \code{F} and \code{M}}
\item{Bwt}{Body weight in kilograms}
\item{Hwt}{Heart weight in kilograms}
}
}
}
\source{
Fisher, R.A. (1947). The analysis of covariance method for the
relation between a part and the whole. Biometrics, 3 2, 65-8.
}
\usage{
subcats
}
\description{
A subset of cats. The heart and body weights of samples of male and female
cats used for digitalis experiments. The cats were all adult, over 2 kg body
weight.
}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics
with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
\keyword{datasets}
