% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_resample}
\alias{summarize_resample}
\title{Summarize Resampled Data}
\usage{
summarize_resample(data, alpha = 0.05)
}
\arguments{
\item{data}{Data.frame of resampled data from \code{\link[=resample]{resample()}}.}

\item{alpha}{Significance level for confidence intervals.}
}
\value{
A dataframe. It has rows for each column in data. The columns are:
\itemize{
\item estimates: Mean of the parameter resampled estimates.
\item sd: Standard deviations of the resampled estimates.
\item lower: Lower confidence interval from empirical quantiles of the
resampled data.
\item upper: Upper confidence interval from empirical quantiles of the
resampled data.
}
}
\description{
Summarize the results of resampled data. Such data is
returned from \code{resample()}, typically when \code{fn()} is given.
}
\examples{
n <- 50
data <- data.frame(
  output = rnorm(n),
  predictor1 = rnorm(n),
  predictor2 = rnorm(n)
)

fn <- function(data) {
  coef(lm(output ~ ., data = data))
}
# Warning, M is below recommendation for example speed
resample_data <- resample(data = data, fn = fn, M = 10, ignore.columns = "output")
results <- summarize_resample(resample_data)
}
