% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Sample Titanic Data}
\format{
\subsection{\code{titanic}}{

A data frame with 714 rows and 5 columns:
\describe{
\item{Survived}{Bernoulli variable (0 or 1) if passenger survived.}
\item{Pclass}{Class of passenger, values of 1,2,3.}
\item{Sex}{Sex of passenger, 'male' or 'female'.}
\item{Age}{Age of passenger}
\item{Parch}{Number of Parents/Children Aboard}
}
}
}
\source{
Hendricks P (2015). \emph{titanic: Titanic Passenger Survival Data Set}.
R package version 0.1.0, \url{https://CRAN.R-project.org/package=titanic}.
}
\usage{
titanic
}
\description{
Data about passengers on the titanic.
}
\keyword{datasets}
