% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tests.R
\name{anova_tests}
\alias{anova_tests}
\title{ANOVA (and similar) Tests for Mean/Median Differences}
\usage{
anova_tests(data, tests = c("anova", "welch", "bayes", "kruskal", "median"))
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}

\item{tests}{Vector of strings, or a string, indicating the tests to check.
Options include 'anova', 'welch', 'bayes', 'kruskal', and 'median'}
}
\value{
A list of means, medians, and anova test results
}
\description{
Compute and return information on tests for the homogeneity of data.
}
\details{
Tests require independent data unless otherwise specified.
\emph{anova}: Classic ANOVA. Requires normal data with homogeneous variance and
similar group sizes
\emph{welch}: ANOVA modification that allows nonhomogeneous variances and is more
robust to varying group sizes
\emph{bayes}: Bayesian approach to group differences
\emph{kruskal}: Also called the Krukal-wallis test. Non-parametric test
\emph{median}: Also called Brown-Mood median test. Non-parametric test with
Bonferroni correction
}
\examples{
data <- data.frame(
  "value" = c(rnorm(14, sd = 2), rnorm(6), rnorm(20, mean = 2)),
  "group" = c(rep("A", 14), rep("B", 6), rep("C", 20))
)
anova_tests(data)
}
\seealso{
\code{\link[stats:aov]{stats::aov()}}, \code{\link[stats:oneway.test]{stats::oneway.test()}}, \code{\link[BayesFactor:anovaBF]{BayesFactor::anovaBF()}},
\code{\link[stats:kruskal.test]{stats::kruskal.test()}}, \code{\link[PMCMRplus:medianTest]{PMCMRplus::medianTest()}}
}
