% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{normality_tests}
\alias{normality_tests}
\title{Normality Tests}
\usage{
normality_tests(
  x,
  tests = c("shapiro", "ks", "ad", "cvm", "lilliefors", "pearson", "sf")
)
}
\arguments{
\item{x}{Vector of numerics to check normality}

\item{tests}{Vector of strings, or a string, indicating the tests to check.
Options include 'shapiro', 'ks', 'ad', 'cvm', 'lilliefors', 'pearson', and
'sf'}
}
\value{
A list of qqplot and normality test results
}
\description{
Compute and return information on tests for the normality of data.
}
\details{
Tests require independent data unless otherwise specified.
\emph{shapiro}: Shapiro-Wilk test
\emph{ks}: Kolmogorov-Smirnov test
\emph{ad}: Anderson-Darling test
\emph{cvm}: Cramer-von Mises test
\emph{lilliefors}: lilliefors test
\emph{pearson}: Pearson chi-square test
\emph{sf}: Shapiro-Francia test
}
\examples{
x <- c(rnorm(14,sd = 2), rnorm(6), rnorm(20,mean=2))
normality_tests(x)
}
\seealso{
\code{\link[stats:shapiro.test]{stats::shapiro.test()}}, \code{\link[stats:ks.test]{stats::ks.test()}}, \code{\link[nortest:ad.test]{nortest::ad.test()}},
\code{\link[nortest:cvm.test]{nortest::cvm.test()}}, \code{\link[nortest:lillie.test]{nortest::lillie.test()}}, \code{\link[nortest:pearson.test]{nortest::pearson.test()}},
\code{\link[nortest:sf.test]{nortest::sf.test()}}
}
