% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{resample_differences}
\alias{resample_differences}
\title{Effect Size by Resampling Differences}
\usage{
resample_differences(data, alpha = 0.05, M = 1000)
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}

\item{alpha}{Significance for confidence intervals, defaults to 0.05}

\item{M}{Numeric. Number of resample iterations.}
}
\value{
Table with the mean differences, confidence intervals and other
information
}
\description{
Use with caution as this is still being tested. This function generates
confidence intervals for the mean differences between groups.
}
\examples{
data <- data.frame('value'=c(rnorm(14,sd = 2), rnorm(6), rnorm(20,mean=2)),
                   'group'=c(rep('A',14),rep('B',6), rep('C',20)))
# Be sure to increase M for real use cases
resample_differences(data, M=50)
}
