% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_distribution.R
\name{resample_distribution}
\alias{resample_distribution}
\title{Resample Distributions}
\usage{
resample_distribution(
  data,
  resampled_data = NULL,
  strata = NULL,
  ignore.columns = NULL,
  ...
)
}
\arguments{
\item{data}{Data.frame (or vector) to be resampled where the rows are the
observations and the columns are the variables. Note, all variables are
permuted not specified in \code{ignore.columns}.}

\item{resampled_data}{Data from \code{\link[=resample]{resample()}} or NULL to run \code{\link[=resample]{resample()}} on
data}

\item{strata}{String or numeric. This indicate the column to stratify the
data when \code{method} is stratify. This can be the column number or the
column name. When NULL the data is not stratified. When given, strata are
sampled separately.}

\item{ignore.columns}{Name or column numbers to ignore when resampling data.
These are not permuted. Note that if less/more samples are collected than
the original, these are permuted separately.}

\item{...}{Additional parameters for \code{\link[=resample]{resample()}}}
}
\value{
ggplot object showing original and resample distribution. Note that
if a function is used, they many not be describing the same thing
}
\description{
Resample Distributions
}
\examples{
resample_distribution(rnorm(100))
resample_distribution(
  data.frame('data'=c(rnorm(100),rnorm(50,mean=10)),
             'strata'=c(rep('A',100),rep('B',50))),
  strata = 'strata' )
resample_distribution(rnorm(100), fn = mean )
}
