% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{resample_function}
\alias{resample_function}
\title{Resample and apply data}
\usage{
resample_function(data, fn, alpha = 0.05, ...)
}
\arguments{
\item{data}{Data.frame to be resampled where the rows are the observations
and the columns are the variables. Note, all variables except the first are
permuted, where the first is used for the models given in fn.}

\item{fn}{Function. Function to use for each permuted sample. The output
should be one of the following. (1) A numeric values (e.g. mean squared
error), (2) A numeric vector (e.g. vector of coefficients), or (3) a
data.frame with a row for each variable and the first column being the values
of interest.#'}

\item{alpha}{Significance level for confidence intervals}

\item{...}{This is for the parameters of the resampling. See parameters of
\code{resample()}. When not given, the defaults are used.}
}
\value{
A list of items. The first is the results as a dataframe. It has the
following columns with a row per variable.
\itemize{
\item estimates: Mean of the parameter estimates from fn over the
resampled trials.
\item sd: Standard deviations of the estimates from \code{fn} over the
resampled trials.
\item lowerNorm: Lower confidence interval assuming normality using alpha.
\item upperNorm: Upper confidence interval assuming normality using alpha.
\item lowerResamp: Lower confidence interval from data using alpha.
\item upperResamp: Upper confidence interval from data using alpha.
}
The second entry is a list of the resampled data and estimates for reference.
\itemize{
\item samples: A list of the resampled data.
\item function_values: A list of the function output values from the
resampled data. Note, this includes those not used for the estimates.
\item estimates: A data frame of the first column of each element in
function_values. The row means are those used in \code{meanParams}.
}
}
\description{
Resample data and compute the related values from some function. This
function is a convenient an extension of \code{resample()} and a similar
function can be constructed using \code{resample()} when
\code{resample_function()} is insufficient. The code permutes every
column except the first ('output'), then applies the function on
the permuted sample with the (non-permuted) output. The output is used
to create bootstrapped estimates.
}
\examples{
n <- 50
data <- data.frame(
  output = rnorm(n),
  predictor1 = rnorm(n),
  predictor2 = rnorm(n)
)

fn <- function(data) { coef(lm(output ~ ., data = data)) }
# Warning, M is below recommendation for example speed
results <- resample_function(data = data, fn = fn, M = 10, method = 'simple')
}
