% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_tests.R
\name{variance_tests}
\alias{variance_tests}
\title{Variance Tests for Homogeneity}
\usage{
variance_tests(
  data,
  tests = c("ratio", "levene", "bartlett", "fligner"),
  ratio.threshold = 4
)
}
\arguments{
\item{data}{Data.frame with the first column the values and the second column
the group names}

\item{tests}{Vector of strings, or a string, indicating the tests to check.
Options include 'ratio', 'levene', 'bartlett', 'fligner', and 'hartley'}

\item{ratio.threshold}{Numeric threshold for comparision of the variance ratio}
}
\value{
Lists of variances and desired variance tests
}
\description{
Compute and return information on tests for the homogeneity of data.
}
\details{
Tests require independent, quantitative data unless otherwise specified.
\emph{Ratio}: Looks at the ratio of largest to smallest variance. Rule of thumb is
that if this ratio is less than 4, no major violations are observed
\emph{Levene}: No other major assumptions
\emph{Bartlett}: Requires normally distributed data and no major outliers
\emph{Fligner}: Also called Fligner-Killeen. Non-parametric test
\emph{Hartley}: Requires normal distribution and equal number of observations.
}
\examples{
data <- data.frame('value'=c(rnorm(14,sd = 2), rnorm(6), rnorm(20,mean=2)),
                   'group'=c(rep('A',14),rep('B',6), rep('C',20)))
variance_tests(data)
}
\seealso{
\code{\link[car:leveneTest]{car::leveneTest()}}, \code{\link[stats:bartlett.test]{stats::bartlett.test()}},
\code{\link[stats:fligner.test]{stats::fligner.test()}}, \code{\link[PMCMRplus:hartleyTest]{PMCMRplus::hartleyTest()}}
}
